5-4: Hunting BUG in GG_Ancestry
There is a logic path in which the Human ancestry is not being
randomized.  The symptom of the BUG is a "can't concatenate None to str"
when trying to construct the database name.

I found the BUG but there's a lot of noise in this commit.  I want the
BUG fix commit to go through seperately.

Changes to be committed:
	modified:   modules/GG_Ancestry.py