2-3: REfactored GG_City NPC randomization
Validation of "list exists and is validated" occurs much lower in the
call chain.  I wanted a dictionary that stored raw(ish) data internally
before it was translated into human readable statements.

I'm about to reorder the "public" methods at the top of the class.

Then I'm going to write in functionality that will randomize a class and
level based on city statistics.

Changes to be committed:
	modified:   modules/GG_City.py