development: Found BUG in GG_Character
GG_Character's call to GG_Ancestry doesn't use the city object.  That
means that any character's ancestry is randomized has nothing to do with
the city config file that's been passed in.

GG_Ancestry needs to be REfactored to take an optional city object and
use the percentages to randomize a race.

Changes to be committed:
	modified:   README.md