GAGU-14: Fixed gg_menu GGGuild class usage
I removed the GGGuild.load() method (because it wasn't necessary)

As of this commit, the GGGuild class is working.
That includes GGGuild class usage of the GGJob class to generate jobs.
The GGJob class isn't currently generating any non-empty strings, but
this is a good milestone.  Things are working as-is in ways I expect
them to be.

Changes to be committed:
	modified:   gamemaster_guidance/gg_menu.py